WIp: concurrency

+++ Surf Boards/Controllers/SurfBoardsController.cs
@@ -13,6 +13,7 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.VisualBasic;
 using Surf_Boards.Core;
 using Microsoft.Data.SqlClient;
+using Surf_Boards.Migrations;
 
 namespace Surf_Boards.Controllers
 {
@@ -29,7 +30,7 @@ namespace Surf_Boards.Controllers
         }
 
         // GET: SurfBoards
-        
+
         public async Task<IActionResult> Index(string SearchString, string sortOrder, string currentFilter, int? pageNumber)
         {
             //Searchning 
@@ -118,6 +119,7 @@ namespace Surf_Boards.Controllers
             }
 
             var surfBoard = await _context.SurfBoard
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (surfBoard == null)
             {
@@ -152,7 +154,7 @@ namespace Surf_Boards.Controllers
                     string wwwRootPath = _hostEnvironment.WebRootPath;
                     string filename = Path.GetFileNameWithoutExtension(surfBoard.ImageFile.FileName);
                     string extension = Path.GetExtension(surfBoard.ImageFile.FileName);
-                    surfBoard.ImageName = filename = filename + "-" + surfBoard.Id +  extension;
+                    surfBoard.ImageName = filename = filename + "-" + surfBoard.Id + extension;
                     string path = Path.Combine(wwwRootPath + "/Images/", filename);
                     using (var fileStream = new FileStream(path, FileMode.Create))
                     {
@@ -176,7 +178,10 @@ namespace Surf_Boards.Controllers
                 return NotFound();
             }
 
-            var surfBoard = await _context.SurfBoard.FindAsync(id);
+            var surfBoard = await _context.SurfBoard
+           .AsNoTracking()
+           .FirstOrDefaultAsync(m => m.Id == id);
+
             if (surfBoard == null)
             {
                 return NotFound();
@@ -189,28 +194,35 @@ namespace Surf_Boards.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(Guid id, [Bind("Id,BoardName,Length,Width,Thickness,Volume,Boardtype,Price,Equipment,ImageName,ImageFile")] SurfBoard surfBoard)
+        public async Task<IActionResult> Edit(Guid id, byte[] rowVersion, [Bind("Id,BoardName,Length,Width,Thickness,Volume,Boardtype,Price,Equipment,ImageName,ImageFile")] SurfBoard surfBoard)
         {
             if (id != surfBoard.Id)
             {
                 return NotFound();
             }
 
+            if (surfBoard == null)
+            {
+                SurfBoard surboard = new SurfBoard();
+                await TryUpdateModelAsync(surboard);
+                ModelState.AddModelError(string.Empty, "unable to save, item deletede by other user");
+                return View(surboard);
+            }
+
             if (ModelState.IsValid)
             {
+                _context.Entry(surfBoard).Property("RowVersion").OriginalValue = rowVersion;
 
                 try
                 {
-                   
-                   
 
                     // Does the user have added a new image? Then we replace the old 
-                    if (surfBoard.ImageFile != null && surfBoard.ImageFile.Length >0)
+                    if (surfBoard.ImageFile != null && surfBoard.ImageFile.Length > 0)
                     {
-                       // It check if there is an image and if so deletes it 
+                        // It check if there is an image and if so deletes it 
                         if (surfBoard.ImageName != null && surfBoard.ImageFile != null)
                         {
-                        string imagePathe = _hostEnvironment.WebRootPath;
+                            string imagePathe = _hostEnvironment.WebRootPath;
                             var imagePath = Path.Combine(imagePathe + "/Images/", surfBoard.ImageName);
                             if (System.IO.File.Exists(imagePath))
                                 System.IO.File.Delete(imagePath);
@@ -236,25 +248,71 @@ namespace Surf_Boards.Controllers
 
                     _context.Update(surfBoard);
                     await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
                 }
-                catch (DbUpdateConcurrencyException)
+                catch (DbUpdateConcurrencyException ex)
                 {
-                    if (!SurfBoardExists(surfBoard.Id))
+                    var exceptionEntry = ex.Entries.Single();
+                    var clientValues = (SurfBoard)exceptionEntry.Entity;
+                    var databaseEntry = exceptionEntry.GetDatabaseValues();
+                    if (databaseEntry == null)
                     {
-                        return NotFound();
+                        ModelState.AddModelError(string.Empty,
+                            "Unable to save changes. The surfboard was deleted by another user.");
                     }
                     else
                     {
-                        throw;
+                        var databaseValues = (SurfBoard)databaseEntry.ToObject();
+
+                        if (databaseValues.BoardName != clientValues.BoardName)
+                        {
+                            ModelState.AddModelError("Name", $"Current value: {databaseValues.BoardName}");
+                        }
+                        if (databaseValues.Length != clientValues.Length)
+                        {
+                            ModelState.AddModelError("lenght", $"Current value: {databaseValues.Length}");
+                        }
+                        if (databaseValues.Width != clientValues.Width)
+                        {
+                            ModelState.AddModelError("Width", $"Current value: {databaseValues.Width}");
+                        }
+                        if (databaseValues.Thickness != clientValues.Thickness)
+                        {
+                            ModelState.AddModelError("Thickness", $"Current value: {databaseValues.Thickness}");
+                        }
+                        if (databaseValues.Volume != clientValues.Volume)
+                        {
+                            ModelState.AddModelError("Volume", $"Current value: {databaseValues.Volume}");
+                        }
+                        if (databaseValues.Boardtype != clientValues.Boardtype)
+                        {
+                            ModelState.AddModelError("Type", $"Current value: {databaseValues.Boardtype}");
+                        }
+                        if (databaseValues.Price != clientValues.Price)
+                        {
+                            ModelState.AddModelError("price", $"Current value: {databaseValues.Price}");
+                        }
+                        if (databaseValues.Equipment != clientValues.Equipment)
+                        {
+                            ModelState.AddModelError("Equipment", $"Current value: {databaseValues.Equipment}");
+                        }
+
+                        ModelState.AddModelError(string.Empty, "The record you attempted to edit "
+                                + "was modified by another user after you got the original value. The "
+                                + "edit operation was canceled and the current values in the database "
+                                + "have been displayed. If you still want to edit this record, click "
+                                + "the Save button again. Otherwise click the Back to List hyperlink.");
+                        surfBoard.RowVersion = (byte[])databaseValues.RowVersion;
+                        ModelState.Remove("RowVersion");
                     }
+
                 }
-                return RedirectToAction(nameof(Index));
             }
             return View(surfBoard);
         }
 
         // GET: SurfBoards/Delete/5
-        public async Task<IActionResult> Delete(Guid? id)
+        public async Task<IActionResult> Delete(Guid? id, bool? concurrencyError)
         {
             if (id == null || _context.SurfBoard == null)
             {
@@ -262,38 +320,56 @@ namespace Surf_Boards.Controllers
             }
 
             var surfBoard = await _context.SurfBoard
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (surfBoard == null)
             {
+                if (concurrencyError.GetValueOrDefault())
+                {
+                    return RedirectToAction(nameof(Index));
+                }
                 return NotFound();
             }
 
+            if (concurrencyError.GetValueOrDefault())
+            {
+                ViewData["ConcurrencyErrorMessage"] = "The record you attempted to delete "
+                    + "was modified by another user after you got the original values. "
+                    + "The delete operation was canceled and the current values in the "
+                    + "database have been displayed. If you still want to delete this "
+                    + "record, click the Delete button again. Otherwise "
+                    + "click the Back to List hyperlink.";
+            }
             return View(surfBoard);
         }
 
         // POST: SurfBoards/Delete/5
-        [HttpPost, ActionName("Delete")]
+        [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(Guid id)
+        public async Task<IActionResult> Delete(SurfBoard surfBoard)
         {
-            if (_context.SurfBoard == null)
+
+            try
             {
-                return Problem("Entity set 'Surf_BoardsContext.SurfBoard'  is null.");
-            }
-            var surfBoard = await _context.SurfBoard.FindAsync(id);
-            if (surfBoard != null)
-            {
-                if (surfBoard.ImageName != null)
+                if (await _context.SurfBoard.AnyAsync(m => m.Id == surfBoard.Id))
                 {
-                    var imagePath = Path.Combine(_hostEnvironment.WebRootPath, "Images", surfBoard.ImageName);
-                    if (System.IO.File.Exists(imagePath))
-                        System.IO.File.Delete(imagePath);
+                    if (surfBoard.ImageName != null)
+                    {
+                        var imagePath = Path.Combine(_hostEnvironment.WebRootPath, "Images", surfBoard.ImageName);
+                        if (System.IO.File.Exists(imagePath))
+                            System.IO.File.Delete(imagePath);
+                    }
+                    _context.SurfBoard.Remove(surfBoard);
+                    await _context.SaveChangesAsync();
                 }
-                _context.SurfBoard.Remove(surfBoard);
+            return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                //Log the error (uncomment ex variable name and write a log.)
+                return RedirectToAction(nameof(Delete), new { concurrencyError = true, id = surfBoard.Id});
             }
 
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
         }
 
         private bool SurfBoardExists(Guid id)
